Remove support for the location tag
The location tag is malleable and therefore potentially attacker-controlled.
Unlike other caveats it can be removed or manipulated by an attacker holding
a Macaroon. This is surprising to me, and something I consider to be an
unnecessarily sharp edge in the format's design.

This change ignores the location field if present in Macaroons, and otherwise
removes it from macaroons::token::Token.